requirements.txt
# VoiceClone Studio - Dependencies
# Python 3.11 recommended

# Core
streamlit>=1.28.0

# Audio processing
torch>=2.0.0
torchaudio>=2.0.0

# TTS Model
chatterbox-tts>=0.1.0

# Optional: for better audio handling
soundfile>=0.12.0
numpy>=1.24.0
